Added comments and reorganized the file structure.
